LayoutTests/fast/security-labels/cookie-event-expected.txt

This tests that the checkcookiewrite event behaves properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


checkcookiewrite reached
PASS tag.isOn(event.securityLabel) is true
checkcookiewrite reached
PASS tag.isOn(event.securityLabel) is false
PASS successfullyParsed is true

TEST COMPLETE
